requirements.txt
django==1.7.1
PyJWT==0.4.2
rsa==3.0
feedparser==5.1.1
requests==2.2.1
pycrypto==2.4
